Fix mobile sdk - renaming
